[css-grid] Compute again the baseline offsets during step 3
The new Baseline Alignment algorithm states that items with sizing
cyclic dependencies must be excluded from any baseline context they
participate in. One of these cyclic dependencies can happen with
intrinsic sized grid areas and relative items.

The grid spec states [1] that flex-sized tracks should be considered as
content-sized when the grid container has an indefinite size. We were
using the AvailableSize(direction) function to determine whether the
grid container is indefinite or not. However, this function doesn't
provide consistent results during the different phases of the grid
layout  logic.

The new Baseline Alignment logic is now integrated in the Grid Track
sizing algorithm. Hence, we need to ensure that an item that
participates in any baseline alignment context during the track sizing
also does during the alignment phase, at the end of the grid layout
logic.

However, the AvailableSize(direction) could indicate that the grid has
an indefinite size during the track sizing, but it returns a definite
size during alignment. This issue leads to inconsistencies regarding
the participation of an item in baseline alignment, so it may raises
the assertion of having a context, like it happens in the bugs this CL
tries to fix.

For inline-axis, we'll assume indefinite constraints only during
intrinsic size computation; afterwards, we'll always assume definite
size. This issue could imply that baseline shims are not considered to
compute the grid's intrinsic size, but are then used to calculate the
actual size of the grid tacks, which may lead to such tracks
overflowing the grid container.

In the case of block-axis sizing, we determine the indefinite or
definite size constraints during layout. However, the presence of
orthogonal grid items imply the execution of the 3rd step of the grid
sizing, repeating the execution of track sizing algorithm in both
directions. During this phase, we can't assume indefinite size
constraints anymore because block-axis has been determined already in
the previous 2 steps. While baseline shims computed in the previous
steps shouldn't vary even if the tracks' size change during this phase,
it's possible that items not participating in baseline alignment before
may do participate in this phase. The inclusion of new items in the
baseline context may indeed change the baseline shims, which will be
used to determine the tracks' size during this step.

This CL removes the early return that prevented he Baseline offsets to
be computed again during the step 3. This change assumes the already
mentioned inconsistencies, which I hope we can clarify in the issue [2]
I filed for the CSS WG.

[1] https://drafts.csswg.org/css-grid/#fr-unit
[2] https://github.com/w3c/csswg-drafts/issues/3046

Bug: 867833, 874861, 876593
Change-Id: I668d399b920c9280a8e20b3e8362f562eded4770